Tolerate NULLs in the 'latest_deletion' field, direct Neutron commands.
In some cases, probably where a project has never had an instance, the 'latest_deletion' field from the `idle_projects` function will be `NULL`/`None`.

To avoid some of the excessive bash gymnastics to cron the job (create a subshell, source the RC file, pipe output in the command to the neutron client), just allow it to consume the RC file and issue the HTTP commands itself.