Remove prometheus pushgateway integration
This integration can break if the IP address of the push gateway ever
changes. We are now using a solution where the hammer is run and then a
separate process on the host will capture the exit status and push a
metric to a push gateway itself, which makes this whole thing a bit
simpler.