Fix region selection in maintenance script
Not sure how this reliably worked before, but the region was not being
specified explicitly to Ironic, and it would try to guess one based on
the session object. It tends to pick the first item returned. Perhaps
this behavior was different before.

The fix is to use an Adapter object, which is like a session but it has
some additional default args that are normally specified at request
time; it just pre-fills those args for you, to e.g. always only look at
a particular region/interface.